requirements.txt
aiogram @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/01/9c/04/c1143501f88479532769c85a3081cb94699c9d4eba1079d3a0e341c8ea/aiogram-2.19-py3-none-any.whl
aiohttp @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/df/5c/c3/19cc08c05b23aaf8a5cf56b108f3a12fa7c3e760219c22ffb1902104f7/aiohttp-3.8.1-cp310-cp310-macosx_11_0_arm64.whl
aiosignal @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/30/30/0d/903be6c7ed45c2c0427a2a18486bb6ceb0a700c346a94414b6b395e16e/aiosignal-1.2.0-py3-none-any.whl
async-timeout @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/5e/15/15/173c5bae6f8d28fa399857d021326b2b05fc2debcbc7bcae7c5bf79891/async_timeout-4.0.2-py3-none-any.whl
asyncpg @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/cd/ca/e8/a99ed8a36f9419b6f90428cde4ad3df291b72806f1c1dbc4c628e33165/asyncpg-0.25.0.tar.gz
attrs @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/c4/ee/75/fcd6c2f4c5a1d60b1bcd97014702ca01b3f032fd927d88780b38708409/attrs-21.4.0-py2.py3-none-any.whl
Babel @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/ed/4e/c7/8c4387255e163b817646635c2c88a7debca2361b2ae43c3a455fca786b/Babel-2.9.1-py2.py3-none-any.whl
certifi @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/89/06/5b/e4bfdf1b8370bdc0a26198e81afa98f68c01a8b1b515c9108c6ad98d8a/certifi-2021.10.8-py2.py3-none-any.whl
charset-normalizer @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/d5/27/31/db4fb74906e3a7f55f720e0079ac1850dd86e30651cdfa5e1f04c53cfa/charset_normalizer-2.0.12-py3-none-any.whl
docopt @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/f9/b1/71/2cc20aece0c959bd3d67ce6ef60d31121a41366c9fbc5be06f000c2bd4/docopt-0.6.2.tar.gz
document @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/db/20/8e/cfd8e9a3727f78fd71f66a1ea81754457049de946df1fea916451bda7c/document-1.0.tar.gz
docx @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/d2/20/cf/eecfc2cdeba8f02084097e2b4898ed07fb61408622479f61916c85002b/docx-0.2.4.tar.gz
docxcompose @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/47/58/a9/df419acb8cbd746b94e605dd333bebe812361f7be0aa885e662c1d0bf6/docxcompose-1.3.4.tar.gz
docxtpl @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/67/e3/24/b408146244694ce0774a006f1c9eb7a985e9c668cb4172193c5585f739/docxtpl-0.15.2-py2.py3-none-any.whl
frozenlist @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/42/b7/75/93bb3d0dd2b0ee826dad5d91380ce0f949354eb9988dbbe38355e4b447/frozenlist-1.3.0-cp310-cp310-macosx_11_0_arm64.whl
idna @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/90/36/8c/81eabf6ac88608721ab27f439c9a6b9a8e6a21cc58c59ebb1a42720199/idna-3.3-py3-none-any.whl
Jinja2 @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/23/8e/7b/3e25069406634e991e0c87bc3d385883dfbd02f33749b43745f63a5cde/Jinja2-3.1.1-py3-none-any.whl
loguru @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/36/1f/fa/e70e76ec49384e6de44a56eedd13608349891af6858a468e0f06d336a5/loguru-0.6.0-py3-none-any.whl
lxml @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/04/35/9f/1686083362b7098c6b7ec5463e4511e9109aae67b152ad50274204b46b/lxml-4.8.0.tar.gz
MarkupSafe @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/17/b2/ca/c418fbb258b15d02aff6db598c5169b95eb81935f5e653b73401ef4f6a/MarkupSafe-2.1.1-cp310-cp310-macosx_10_9_universal2.whl
multidict @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/7c/69/c9/ef7bab4d56e743954573abc690127c3552a69b89fbc053c241b216869d/multidict-6.0.2-cp310-cp310-macosx_11_0_arm64.whl
num2words @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/be/33/c1/ac2ab56b57c46ef96f65030d85872ced860c82e9e25ebc01a2e8f298d4/num2words-0.5.10-py3-none-any.whl
Pillow @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/8f/48/99/0d5ba468ebbabf5a3ff1726d194ddbae36a21577026ec110d0d283f5cb/Pillow-9.1.0-cp310-cp310-macosx_11_0_arm64.whl
python-docx @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/15/23/08/1cd747f1bb2b028d37bf196e993280d4e1a1087e36784dd2c34867fbdf/python-docx-0.8.11.tar.gz
pytz @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/76/c4/54/e82c514018d400ab72bc3cc53469bd3b519f30e674c7d905fd7c872314/pytz-2022.1-py2.py3-none-any.whl
requests @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/d2/b2/c6/a04ce59140c6739203837d8dd0f518e29051b7ab61d2f34d4fd4241d30/requests-2.27.1-py2.py3-none-any.whl
six @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/89/b2/f8/fd92b6d5daa0f8889429b2fc67ec21eedc5cae5d531ee2853828ced6c7/six-1.16.0-py2.py3-none-any.whl
urllib3 @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/88/1c/d5/a55ed0245e5d7cd3a9f40dd75733644cbf7b7d94a6c521eb6c027a326c/urllib3-1.26.9-py2.py3-none-any.whl
yadisk @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/8e/a3/4d/31c241b371dc6af0db44967249143383aceb2fbd108fcadfae1bb17f7e/yadisk-1.2.15-py3-none-any.whl
yarl @ file:///Users/stepanmartynov/Library/Caches/pypoetry/artifacts/0f/c5/a5/c00a4615b1af40c84c7f43bb7cac4f3c244d6bcb66aff26ca8144ef596/yarl-1.7.2-cp310-cp310-macosx_11_0_arm64.whl
